Many Changes: Replaced Text View widget to visualize the scanned barcodes in favor of a List View widget. Moved the storage if scouting records to a more OOP style. Made a few new classes for ease of implementation.
Next Changes:
Extended summary mode when a list item is tapped on to give the option to delete the record or maybe tweak the data.

Eventually hope to add the possibility to select the output file and location to support keeping more than one file on the device.

Will eventually remove the trash button on Main Screen and change the color scheme.