Many Changes: Replaced Text View widget to visualize the scanned barcodes in favor of a List View widget. Moved the storage if scouting records to a more OOP style. Made a few new classes for ease of implementation.
Next Changes:
Extended summary mode when a list item is tapped on to give the option to delete the record or maybe tweak the data.

Eventually hope to add the possibility to select the output file and location to support keeping more than one file on the device.

Will eventually remove the trash button on Main Screen and change the color scheme.

Hoping to add functionality for bluetooth keyboard emulating barcode scanner for more device compatibility

***************************************************************************
This version removes the delete all button, and moves it to a per record basis
The list view now has functionality when a list item is taped, it will open a summary page with a delete button and info about the scanned data.
Added settings image button, but doesn't function at this time.

***************************************************************************
Added text animation to the Scan QR Code instruction, and added debug feature to log all scanned codes to a file.

***************************************************************************
Implemented a HashMap to combat the issue of a missing field on the input string.